OPE-9 Extract tool validation graph node
