+++ src/opentulpa/agent/graph_nodes/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime graph node factories."""
